+++ bt-core/src/main/java/bt/data/Bitfield.java
@@ -48,7 +48,13 @@ public class Bitfield {
      * Bitmask indicating availability of pieces.
      * If the n-th bit is set, then the n-th piece is complete and verified.
      */
-    private final BitSet value;
+    private final BitSet bitmask;
+
+    /**
+     * Bitmask indicating pieces that should be skipped.
+     * If the n-th bit is set, then the n-th piece should be skipped.
+     */
+    private volatile BitSet skipped;
 
     /**
      * Total number of pieces in torrent.
@@ -71,7 +77,7 @@ public class Bitfield {
      */
     public Bitfield(List<ChunkDescriptor> chunks) {
         this.piecesTotal = chunks.size();
-        this.value = new BitSet(chunks.size());
+        this.bitmask = new BitSet(chunks.size());
         this.chunks = Optional.of(chunks);
         this.lock = new ReentrantLock();
     }
@@ -85,7 +91,7 @@ public class Bitfield {
      */
     public Bitfield(int piecesTotal) {
         this.piecesTotal = piecesTotal;
-        this.value = new BitSet(piecesTotal);
+        this.bitmask = new BitSet(piecesTotal);
         this.chunks = Optional.empty();
         this.lock = new ReentrantLock();
     }
@@ -118,7 +124,7 @@ public class Bitfield {
      */
     public Bitfield(byte[] value, BitOrder bitOrder, int piecesTotal) {
         this.piecesTotal = piecesTotal;
-        this.value = createBitmask(value, bitOrder, piecesTotal);
+        this.bitmask = createBitmask(value, bitOrder, piecesTotal);
         this.chunks = Optional.empty();
         this.lock = new ReentrantLock();
     }
@@ -156,7 +162,7 @@ public class Bitfield {
     public BitSet getBitmask() {
         lock.lock();
         try {
-            return Protocols.copyOf(value);
+            return Protocols.copyOf(bitmask);
         } finally {
             lock.unlock();
         }
@@ -175,8 +181,8 @@ public class Bitfield {
 
         lock.lock();
         try {
-            bytes = value.toByteArray();
-            truncated = (value.length() < piecesTotal);
+            bytes = bitmask.toByteArray();
+            truncated = (bitmask.length() < piecesTotal);
         } finally {
             lock.unlock();
         }
@@ -206,16 +212,81 @@ public class Bitfield {
      * @since 1.0
      */
     public int getPiecesComplete() {
-        return value.cardinality();
+        lock.lock();
+        try {
+            return bitmask.cardinality();
+        } finally {
+            lock.unlock();
+        }
     }
 
     /**
-     * @return Number of pieces that have status different from {@link PieceStatus#COMPLETE_VERIFIED}.
-     *         I.e. it's the same as {@link #getPiecesTotal()} - {@link #getPiecesComplete()}
+     * @return Number of pieces that have status different {@link PieceStatus#COMPLETE_VERIFIED}.
+     * @since 1.7
+     */
+    public int getPiecesIncomplete() {
+        lock.lock();
+        try {
+            return getPiecesTotal() - bitmask.cardinality();
+        } finally {
+            lock.unlock();
+        }
+    }
+
+    /**
+     * @return Number of pieces that have status different from {@link PieceStatus#COMPLETE_VERIFIED}
+     *         and should NOT be skipped.
      * @since 1.0
      */
     public int getPiecesRemaining() {
-        return getPiecesTotal() - getPiecesComplete();
+        lock.lock();
+        try {
+            if (skipped == null) {
+                return getPiecesTotal() - getPiecesComplete();
+            } else {
+                BitSet bitmask = getBitmask();
+                bitmask.or(skipped);
+                return getPiecesTotal() - bitmask.cardinality();
+            }
+        } finally {
+            lock.unlock();
+        }
+    }
+
+    /**
+     * @return Number of pieces that should be skipped
+     * @since 1.7
+     */
+    public int getPiecesSkipped() {
+        if (skipped == null) {
+            return 0;
+        }
+
+        lock.lock();
+        try {
+            BitSet bitmask = Protocols.copyOf(skipped);
+            bitmask.andNot(this.bitmask);
+            return bitmask.cardinality();
+        } finally {
+            lock.unlock();
+        }
+    }
+
+    /**
+     * @return Number of pieces that should NOT be skipped
+     * @since 1.7
+     */
+    public int getPiecesNotSkipped() {
+        if (skipped == null) {
+            return piecesTotal;
+        }
+
+        lock.lock();
+        try {
+            return piecesTotal - skipped.cardinality();
+        } finally {
+            lock.unlock();
+        }
     }
 
     /**
@@ -232,7 +303,7 @@ public class Bitfield {
         boolean verified;
         lock.lock();
         try {
-            verified = value.get(pieceIndex);
+            verified = this.bitmask.get(pieceIndex);
         } finally {
             lock.unlock();
         }
@@ -289,7 +360,7 @@ public class Bitfield {
 
         lock.lock();
         try {
-            value.set(pieceIndex);
+            bitmask.set(pieceIndex);
         } finally {
             lock.unlock();
         }
@@ -307,7 +378,45 @@ public class Bitfield {
 
     private void validatePieceIndex(Integer pieceIndex) {
         if (pieceIndex < 0 || pieceIndex >= getPiecesTotal()) {
-            throw new BtException("Illegal piece index: " + pieceIndex);
+            throw new BtException("Illegal piece index: " + pieceIndex +
+                    ", expected 0.." + (getPiecesTotal() - 1));
+        }
+    }
+
+    /**
+     * Mark a piece as skipped
+     *
+     * @since 1.7
+     */
+    public void skip(int pieceIndex) {
+        validatePieceIndex(pieceIndex);
+
+        lock.lock();
+        try {
+            if (skipped == null) {
+                skipped = new BitSet(getPiecesTotal());
+            }
+            skipped.set(pieceIndex);
+        } finally {
+            lock.unlock();
+        }
+    }
+
+    /**
+     * Mark a piece as not skipped
+     *
+     * @since 1.7
+     */
+    public void unskip(int pieceIndex) {
+        validatePieceIndex(pieceIndex);
+
+        if (skipped != null) {
+            lock.lock();
+            try {
+                skipped.clear(pieceIndex);
+            } finally {
+                lock.unlock();
+            }
         }
     }
 }