+++ bt-core/src/main/java/bt/data/DataDescriptor.java
@@ -16,6 +16,8 @@
 
 package bt.data;
 
+import bt.metainfo.TorrentFile;
+
 import java.io.Closeable;
 import java.util.List;
 
@@ -40,4 +42,12 @@ public interface DataDescriptor extends Closeable {
      * @since 1.0
      */
     Bitfield getBitfield();
+
+    /**
+     * Get a list of files that a given piece index intersects
+     *
+     * @return A list of files that a given piece index intersects
+     * @since 1.7
+     */
+    List<TorrentFile> getFilesForPiece(int pieceIndex);
 }