+++ bt-core/src/main/java/bt/processor/torrent/InitializeTorrentProcessingStage.java
@@ -19,8 +19,8 @@ package bt.processor.torrent;
 import bt.data.Bitfield;
 import bt.event.EventSink;
 import bt.metainfo.Torrent;
-import bt.processor.TerminateOnErrorProcessingStage;
 import bt.processor.ProcessingStage;
+import bt.processor.TerminateOnErrorProcessingStage;
 import bt.processor.listener.ProcessingEvent;
 import bt.runtime.Config;
 import bt.torrent.BitfieldBasedStatistics;
@@ -28,18 +28,12 @@ import bt.torrent.TorrentDescriptor;
 import bt.torrent.TorrentRegistry;
 import bt.torrent.data.DataWorker;
 import bt.torrent.data.IDataWorkerFactory;
-import bt.torrent.messaging.Assignments;
 import bt.torrent.messaging.BitfieldConsumer;
 import bt.torrent.messaging.GenericConsumer;
-import bt.torrent.messaging.IncompletePiecesValidator;
 import bt.torrent.messaging.MetadataProducer;
 import bt.torrent.messaging.PeerRequestConsumer;
 import bt.torrent.messaging.PieceConsumer;
 import bt.torrent.messaging.RequestProducer;
-import bt.torrent.selector.PieceSelector;
-import bt.torrent.selector.ValidatingSelector;
-
-import java.util.function.Predicate;
 
 public class InitializeTorrentProcessingStage<C extends TorrentContext> extends TerminateOnErrorProcessingStage<C> {
 
@@ -67,10 +61,8 @@ public class InitializeTorrentProcessingStage<C extends TorrentContext> extends 
 
         Bitfield bitfield = descriptor.getDataDescriptor().getBitfield();
         BitfieldBasedStatistics pieceStatistics = createPieceStatistics(bitfield);
-        PieceSelector selector = createSelector(context.getPieceSelector(), bitfield);
 
         DataWorker dataWorker = createDataWorker(descriptor);
-        Assignments assignments = new Assignments(bitfield, selector, pieceStatistics, config);
 
         context.getRouter().registerMessagingAgent(GenericConsumer.consumer());
         context.getRouter().registerMessagingAgent(new BitfieldConsumer(bitfield, pieceStatistics, eventSink));
@@ -80,18 +72,11 @@ public class InitializeTorrentProcessingStage<C extends TorrentContext> extends 
         context.getRouter().registerMessagingAgent(new MetadataProducer(() -> context.getTorrent().orElse(null), config));
 
         context.setBitfield(bitfield);
-        context.setAssignments(assignments);
         context.setPieceStatistics(pieceStatistics);
     }
 
     private BitfieldBasedStatistics createPieceStatistics(Bitfield bitfield) {
         return new BitfieldBasedStatistics(bitfield);
-    }
-
-    private PieceSelector createSelector(PieceSelector selector,
-                                         Bitfield bitfield) {
-        Predicate<Integer> validator = new IncompletePiecesValidator(bitfield);
-        return new ValidatingSelector(validator, selector);
     }
 
     private DataWorker createDataWorker(TorrentDescriptor descriptor) {